refactor: reorganize files to separate concerns and be explicit about exports
