+++ test/SgidClient.spec.ts
@@ -1,7 +1,11 @@
 import { readFileSync } from 'fs'
 
 import SgidClient from '../src'
-import { codeVerifierAndChallengePattern } from '../src/util'
+import {
+  generateCodeChallenge,
+  generateCodeVerifier,
+  generatePkcePair,
+} from '../src/generators'
 
 import {
   MOCK_ACCESS_TOKEN,
@@ -24,6 +28,7 @@ import {
   userInfoHandlerNoData,
   userInfoHandlerNoKey,
 } from './mocks/handlers'
+import { codeVerifierAndChallengePattern } from './mocks/helpers'
 import { server } from './mocks/server'
 
 /**
@@ -391,7 +396,7 @@ describe('SgidClient', () => {
 
   describe('generatePkcePair', () => {
     it('should generate a PKCE pair when no length is provided', () => {
-      const pkcePair = SgidClient.generatePkcePair()
+      const pkcePair = generatePkcePair()
 
       expect(pkcePair.codeVerifier.length).toBe(43)
       expect(pkcePair.codeChallenge.length).toBe(43)
@@ -402,7 +407,7 @@ describe('SgidClient', () => {
 
     it('should generate a PKCE pair of specified length when length between 43 (inclusive) and 128 (inclusive) is provided', () => {
       for (let length = 43; length <= 128; length++) {
-        const pkcePair = SgidClient.generatePkcePair(length)
+        const pkcePair = generatePkcePair(length)
 
         // Length is only for the code verifier
         expect(pkcePair.codeVerifier.length).toBe(length)
@@ -414,7 +419,7 @@ describe('SgidClient', () => {
 
     it('should throw an error when a length < 43 or length > 128 is provided', () => {
       for (const length of [-1, 0, 42, 129, 138, 999]) {
-        expect(() => SgidClient.generatePkcePair(length)).toThrowError(
+        expect(() => generatePkcePair(length)).toThrowError(
           'The function generatePkcePair should receive a minimum length of 43 and a maximum length of 128',
         )
       }
@@ -423,7 +428,7 @@ describe('SgidClient', () => {
 
   describe('generateCodeVerifier', () => {
     it('should generate a code verifier of length 43 when no length is provided', () => {
-      const codeVerifier = SgidClient.generateCodeVerifier()
+      const codeVerifier = generateCodeVerifier()
 
       expect(codeVerifier.length).toBe(43)
       expect(codeVerifier).toMatch(codeVerifierAndChallengePattern)
@@ -431,7 +436,7 @@ describe('SgidClient', () => {
 
     it('should generate a code verifier of specified length when length between 43 (inclusive) and 128 (inclusive) is provided', () => {
       for (let length = 43; length <= 128; length++) {
-        const codeVerifier = SgidClient.generateCodeVerifier(length)
+        const codeVerifier = generateCodeVerifier(length)
         expect(codeVerifier.length).toBe(length)
         expect(codeVerifier).toMatch(codeVerifierAndChallengePattern)
       }
@@ -439,7 +444,7 @@ describe('SgidClient', () => {
 
     it('should throw an error when a length < 43 or length > 128 is provided', () => {
       for (const length of [-1, 0, 42, 129, 138, 999]) {
-        expect(() => SgidClient.generateCodeVerifier(length)).toThrowError(
+        expect(() => generateCodeVerifier(length)).toThrowError(
           `The code verifier should have a minimum length of 43 and a maximum length of 128`,
         )
       }
@@ -448,16 +453,14 @@ describe('SgidClient', () => {
 
   describe('generateCodeChallenge', () => {
     it('should match the specified pattern', () => {
-      expect(SgidClient.generateCodeChallenge(MOCK_CODE_VERIFIER)).toMatch(
+      expect(generateCodeChallenge(MOCK_CODE_VERIFIER)).toMatch(
         codeVerifierAndChallengePattern,
       )
     })
 
     it('should be deterministic (return the same code challenge given the same code verifier)', () => {
-      const firstCodeChallenge =
-        SgidClient.generateCodeChallenge(MOCK_CODE_VERIFIER)
-      const secondCodeChallenge =
-        SgidClient.generateCodeChallenge(MOCK_CODE_VERIFIER)
+      const firstCodeChallenge = generateCodeChallenge(MOCK_CODE_VERIFIER)
+      const secondCodeChallenge = generateCodeChallenge(MOCK_CODE_VERIFIER)
 
       expect(firstCodeChallenge).toBe(secondCodeChallenge)
     })
